requirements.txt
langchain 
gradio 
chromadb
unstructured==0.6.8
unstructured-inference==0.4.4
pypdf
google-cloud-aiplatform
tabulate
pyyaml
openai